requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
redis
psycopg2-binary
python-dotenv
certifi
requests
pdfplumber
pytesseract
Pillow
opencv-python-headless
numpy
joblib